+++ Source/DataStructs/Make.package
@@ -9,3 +9,4 @@ CEXE_headers += ERF_SpongeStruct.H
 CEXE_headers += ERF_TurbStruct.H
 CEXE_headers += ERF_TurbPertStruct.H
 CEXE_headers += ERF_LargeScaleForcingData.H
+CEXE_headers += ERF_ShocTransportStruct.H
